backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
openpyxl
pandas
numpy
rank-bm25
python-dotenv
sentence-transformers
faiss-cpu
requests
deep-translator
langdetect
